feat: Allow concurrent execution of the automations
To stop errors from surfacing when the automation attempts to run at the
same time as a previous trigger of the automation, allow parallel
execution. This shouldn't be an issue since all the values used are
calculated by data collected in the execution cycle from sources that
should be shared between the runs.